complete

+++ python_introduction/finance_calculator.py
@@ -5,6 +5,6 @@ total_monthly_expenses = int(input("Enter your total monthly expenses:"))
 annual_interest_rate = 0.05
 monthly_savings = monthly_income - total_monthly_expenses
 
-Projected_savings = monthly_savings * 12 + ( monthly_savings * 12 * 0.05)
+Projected_savings = float(monthly_savings * 12 + ( monthly_savings * 12 * 0.05))
 
 print ( "Projected savings after one year, with interest, is: $", Projected_savings ,".")
